Version 3.31.33 (based on 45a36948e1bb0f02d395ecbcfe6306553898268b)
Performance and stability improvements on all platforms.

Cr-Commit-Position: refs/heads/candidates@{#25307}

+++ src/elements.h
@@ -179,7 +179,7 @@ class ElementsAccessor {
 
   // Returns a shared ElementsAccessor for the specified ElementsKind.
   static ElementsAccessor* ForKind(ElementsKind elements_kind) {
-    DCHECK(elements_kind < kElementsKindCount);
+    DCHECK(static_cast<int>(elements_kind) < kElementsKindCount);
     return elements_accessors_[elements_kind];
   }
